update db.test.ts to seed.test.ts
