+++ packages/cli/test/commands/console.test.ts
@@ -1,8 +1,6 @@
 import {Console} from "../../src/commands/console"
-import * as path from "path"
 
 import * as repl from "@blitzjs/repl"
-import * as db from "../../src/commands/db"
 
 jest.spyOn(global.console, "log").mockImplementation()
 
@@ -37,8 +35,6 @@ jest.mock(
   }),
 )
 
-Console.prototype.parse = jest.fn()
-
 describe("Console command", () => {
   beforeEach(() => {
     jest.resetAllMocks()
@@ -46,13 +42,11 @@ describe("Console command", () => {
 
   it("runs repl", async () => {
     await Console.run()
-    expect(Console.prototype.parse).toHaveBeenCalled()
     expect(repl.runRepl).toHaveBeenCalled()
   })
 
   it("runs repl with replOptions", async () => {
     await Console.run()
-    expect(Console.prototype.parse).toHaveBeenCalled()
     expect(repl.runRepl).toHaveBeenCalledWith(Console.replOptions)
   })
 })
